osutil: move build_id to osutil, scan all section
